pyproject.toml
[project]
name = "function-store-mcp"
version = "0.1.0"
description = "Draft-First Function Store MCP Server"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "duckdb",
    "psutil",
    "tqdm>=4.66.0",
    "ruff>=0.1.0",
    "fastapi",
    "uvicorn",
    "httpx",
    "google-genai",
    "bandit",
    "safety",
    "python-dotenv",
    "gitpython",
    "PyGithub",
]

[project.scripts]
mcp-server-function-store = "edge.mcp_server:main"
logic-hive-setup = "dev_tools.setup_mcp:main"

[tool.setuptools.packages.find]
where = ["backend"]
include = ["core", "hub", "edge"]

[tool.pytest.ini_options]
testpaths = ["dev_tools/tests"]
pythonpath = ["backend"]
asyncio_mode = "auto"
addopts = "--basetemp=dev_tools/tests/test_data"